Implementation of Open Api 3 standard.
Code refactoring:
- support for swagger 1 removed,
- new way for determining swagger version,
- getting rid of unused options (lint, beautify, moduleName),
- splitting functionality into multiple files,
- typescript methods now distinguish between query parameters and body
