+++ lib/cli.js
@@ -1,12 +1,8 @@
-'use strict';
-
-const fs = require('fs');
 const pkg = require('../package.json');
 const cli = require('commander');
-const yaml = require('js-yaml').safeLoad;
 const CodeGen = require('./codegen').CodeGen;
 const fetch = require('isomorphic-fetch');
-const utilities = require('./utilities');
+const utilities = require('./utilities').Utilities;
 
 cli
   .version(pkg.version)
@@ -30,22 +26,40 @@ cli
     const fn = CodeGen[fnName];
 
     fetch(swaggerUrl)
-      .then(response => response.json())
+      .then(response => {
+        if (response.ok) {
+          return response.json();
+        }
+        try {
+          const json = response.json();
+          throw new Error(JSON.stringify(json));
+        } catch (e) {
+          throw new Error('Invalid Swaggger url');
+        }
+      })
       .then(swaggerJson => {
+        const version = utilities.getVersion(swaggerJson);
+        if (![2, 3].includes(version)) {
+          throw 'Only Swagger versions 2 or 3 are supported!';
+        }
+
         const swagger = utilities.filterSwaggerPathsByTags(
           swaggerJson,
           options.tags,
         );
 
-        const result = fn({
-          moduleName: options.module,
-          className: options.class,
-          swagger: swagger,
-          lint: options.lint,
-          beautify: options.beautify,
-        });
+        const result = fn(
+          {
+            className: options.class,
+            swagger: swagger,
+          },
+          version,
+        );
 
         console.log(result);
+      })
+      .catch(e => {
+        console.error(e);
       });
   });
