+++ lib/typescript.js
@@ -15,7 +15,7 @@ function convertType(swaggerType, swagger) {
   var typespec = { description: swaggerType.description, isEnum: false };
 
   if (swaggerType.hasOwnProperty('schema')) {
-    return convertType(swaggerType.schema);
+    return convertType(swaggerType.schema, swagger);
   } else if (_.isString(swaggerType.$ref)) {
     typespec.tsType = 'ref';
     typespec.target = swaggerType.$ref.substring(
@@ -37,9 +37,16 @@ function convertType(swaggerType, swagger) {
     typespec.tsType = 'boolean';
   } else if (swaggerType.type === 'array') {
     typespec.tsType = 'array';
-    typespec.elementType = convertType(swaggerType.items);
+    typespec.elementType = convertType(swaggerType.items, swagger);
   } else if (Array.isArray(swaggerType.type)) {
-    typespec.tsType = swaggerType.type.join('|');
+    typespec.tsType = swaggerType.type
+      .map(t => {
+        if (t === 'number' || t === 'integer') {
+          return 'number';
+        }
+        return t;
+      })
+      .join('|');
     typespec.isAtomic = true;
     typespec.isEnum = true;
   } else {
@@ -73,7 +80,7 @@ function convertType(swaggerType, swagger) {
               }
             });
           } else {
-            var property = convertType(ref);
+            var property = convertType(ref, swagger);
             Array.prototype.push.apply(
               typespec.properties,
               property.properties,
@@ -83,7 +90,7 @@ function convertType(swaggerType, swagger) {
       }
 
       _.forEach(swaggerType.properties, function(propertyType, propertyName) {
-        var property = convertType(propertyType);
+        var property = convertType(propertyType, swagger);
         property.name = propertyName;
 
         property.optional = true;
