+++ lib/utilities.js
@@ -11,6 +11,21 @@ function filterTagsF(rule, str) {
   return matchRuleShort(rule, str);
 }
 
+function getVersion(swaggerDefinition) {
+  if (swaggerDefinition.swagger === '2.0') {
+    return 2;
+  }
+  if (
+    swaggerDefinition.openapi &&
+    typeof swaggerDefinition.openapi === 'string'
+  ) {
+    if (swaggerDefinition.openapi.match(/3.\d+.\d+/) !== null) {
+      return 3;
+    }
+  }
+  return null;
+}
+
 function filterSwaggerPathsByTags(swagger, tagsToFilter) {
   const operationIds = Object.keys(swagger.paths)
     .reduce((opids, path) => {
@@ -58,4 +73,35 @@ function filterSwaggerPathsByTags(swagger, tagsToFilter) {
   });
 }
 
-exports.filterSwaggerPathsByTags = filterSwaggerPathsByTags;
+var normalizeName = function(id) {
+  return id.replace(/\.|\-|\{|\}|\s/g, '_');
+};
+
+var getPathToMethodName = function(opts, m, path) {
+  if (path === '/' || path === '') {
+    return m;
+  }
+
+  // clean url path for requests ending with '/'
+  var cleanPath = path.replace(/\/$/, '');
+
+  var segments = cleanPath.split('/').slice(1);
+  segments = _.transform(segments, function(result, segment) {
+    if (segment[0] === '{' && segment[segment.length - 1] === '}') {
+      segment =
+        'by' +
+        segment[1].toUpperCase() +
+        segment.substring(2, segment.length - 1);
+    }
+    result.push(segment);
+  });
+  var result = _.camelCase(segments.join('-'));
+  return m.toLowerCase() + result[0].toUpperCase() + result.substring(1);
+};
+
+exports.Utilities = {
+  getVersion,
+  filterSwaggerPathsByTags,
+  normalizeName,
+  getPathToMethodName,
+};
